From: Patrick Ohly <patrick.ohly@intel.com>
The test already covered an EXDATE with TZID. This patch
adds testing with an EXDATE that uses UTZ (Z suffix).

EXDATES in UTC are not particularly useful for events
with TZID (if the time zone definition changes, the
exception might no longer match), but this has been
used for a while by ScheduleWorld.