Remove a test that uses icaltime_as_timet_with_zone to perform a nonsensical transformation. It's asking for a timet for a time by passing in a timezone that the time isn't in. This results in tt2 being different from tt by the GMT offset of the timezone being tested.
This change occurred because of a fix to icaltime_convert_to_zone. As far as I can tell this test was incorrect, and the incorrect behavior of icaltime_convert_to_zone was masking the problem. Other comments in the headers suggest that icaltime_convert_to_zone should behave like it does now that it is fixed, which would imply that this test is wrong.

A better long-term solution would be to remove all the icaltime methods that let you pass in a timezone separate from the zone in the icaltimetype struct. It doesn't make sense to ask libical to perform time operations with some arbitrary timezone that's not attached to the icaltimetype being passed in.