ical_file.cmake: Fix ICAL_CHECK_VERSION macro
The inclusion of the icalversion.h header in ical.h breaks the
ICAL_CHECK_VERSION macro because all the \ line continuation characters
are converted to ;. This makes the macro useless. Wrapping for the sake
of readability, it currently looks like this:

  #define ICAL_CHECK_VERSION(major,minor,micro) ; (ICAL_MAJOR_VERSION >
  (major) || ; (ICAL_MAJOR_VERSION == (major) && ICAL_MINOR_VERSION >
  (minor)) || ; (ICAL_MAJOR_VERSION == (major) && ICAL_MINOR_VERSION ==
  (minor) && ; ICAL_MICRO_VERSION >= (micro)))

Setting the STRINGS NEWLINE_CONSUME option keeps the original structure
instead of terminating the lines.