libical-glib documentation doesn't go through gtkdoc-scangobj
Using gtkdoc-scangobj adds also glib GObject hierarchy, properties and
signals into the documentation. This adds also an ENABLE_GTK_DOC option
which can be used to disable the gtk-doc usage explicitly, rather than
ignore the build when necessary binaries are not found. It also extracts
the commands into a GtkDoc.cmake module, rather than have it inline
on the place where it's used.