[libical-glib] Add some _take_ functions for the C API
Those are:
- i_cal_component_take_property
- i_cal_component_take_component
- i_cal_property_take_parameter
- i_cal_property_take_value
and they just call their _add_ counterparts (and i_cal_property_set_value),
but also unref the argument, thus the corresponding object is properly freed.
This is useful in the C API to avoid unnecessary round trips with the object
creation, add/set it into the other object and unref it. In other words,
constructions like the below are possible with these new functions without
leaking the memory of the property object:

   i_cal_component_take_property (icomp, i_cal_property_new_summary ("Summary"));