Change the icalarray wrapper and update the array.py test
The libical's icalarray is not suitable for introspection much,
because it operates on bare structures, thus rather not allow
to add/get raw objects to/from it and let it be used for the timezone
arrays only.

The test itself could fail under libasan or a memory issue could
be reported when it had been run under valgrind.