i-cal-attach.xml, icalattach.c - improve memory handling
The problem here is the change of icalattach structure, or
better due to change of icalattach_new_from_data(). The
previous versions ignored the 'free' function and always made
a copy of the passed-in data, while libical 3.0.7 doesn't do
the copy and uses the 'free' function. The libical-glib
wrapper for this function makes a copy when no 'free'
function is passed in, which mimics the previous behavior.

Could the icalattach_new_from_data() itself be changed, to
have a similar code as its libical-glib counterpart? That is,
when no 'free' function is passed in, then copy the 'data'
and set the 'free' function on itself? That would preserve
previous behavior.

The attached is a fix for libical doing just that,
which preserves backward compatibility.