libical-glib: Simplify memory management around ICalCompIter
Previously, before this change, the components returned from
the ICalCompIter structures did not hold the reference to the parent
component, thus the parent component could be freed before the returned
instance, effectively destroying the internal libical native structure.
There was required to set the owner manually before, but it breaks
abstraction and is easy to forget.

This change makes all these things transparent. The old code does not need
to be changed, because it should set the same owner anyway, thus the call
results in no change on the ownership of the returned components.