icalvalue: Reset non-UTC icaltimetype::zone on set
When setting an icaltimetype value, the passed-in structure can
have set a timezone, which is tight to the source component
of the passed-in value and which can be made invalid when the source
component is freed. As the structure is copied completely even
on clone of the component, then the 'zone' pointer can be passed
forward in the clones.

Related to https://bugzilla.redhat.com/show_bug.cgi?id=2038090