Add icalparser_ctrl to handle CONTROL characters during parse
RFC 5545 section 3.1 defines CONTROL characters to be byte
values in the decimal ranges [0-8], [10-31] as well as 127.
Of these, all but 10 (LF) and 13 (CR) are not allowed to
be part of any valid content-line produced by the
"contentline" ABNF defined in this section.

However, the current libical parser allows such illegal values
to occur at almost all places within iCalendar and preserves
them in TEXT values as well as non-standard property names.
This causes libical to silently accept invalid iCalendar input
and generate invalid output for such data.

This patch adds the icalparser_ctrl setting to define how to
handle invalid CONTROL characters during parse:

* KEEP: preserve current libical behaviour
* OMIT: silently ignores such characters during parse
* ERROR: rejects the content-line and inserts an error

For backwards-compability, the default setting is KEEP.