icaltime.c - in icaltime_timegm() use make_time() to validate struct tm
In icaltime_timegm(), validate the input struct tm using make_time()
before attempting to convert the structure values to an icaltime_t.

This solves the problem, for example, that tm->tm_mon might
be out-of-range for the days_in_year_passed_month array.

This patch includes a test fuzzer that uncovered this crash.
(see https://issues.oss-fuzz.com/issues/4253856)