icaltypes.c: fix bug ignoring non-empty REQUEST-STATUS extdata
Due to an erroneous conditional, a non-empty 'extdata' part
of a REQUEST-STATUS property value did not get parsed into
the 'debug' field of the icalreqstattype struct type.

It also fixes the docstring in the header that mistakenly
stated that the 'debug' field value points into the original
input string. Instead, it points into the ring buffer memory.