mkderivedparameters.pl: refactor multi-valued parameters API
Refactors the "new_foo" constructor for multi-valued parametes to
take a single element, rather than a list of elements.

Adds the "new_foo_list" constructor to accept a list of elements.

Adds the convenience getter "get_foo_nth" to get the nth value
of a multi-valued parameter. Adds the new "get_foo_size" function
to return the length of the multi-value list.