icalcomponent.c: fix github issue630
RFC 5545 specifies that when DTEND is calcuated from DURATION, the date
part of DURATION (days and weeks) is to be treated as nominal, and the
time part as exact.
Although we allow an event's duration to be calculated from DTSTART and
DTEND, the RFC does not specify how this should be done. We therefore
attempt to reverse the calculation of DTEND from DURATION, as this is
likely what users expect.