icalrecur.c: fix github issues 945 and 787
When iterating over a set of occurrences, we store the previous instance
of the iterator before moving to the next one. If the next instance is
invalid, we restore the previous instance. We also check whether the
previous instance has the same time as the new instance, and if it does
we skip the new instance.