libicalvcard API fixes (#994)
* vcardderivedvalue.h.in: Correct enum type name

Typo in the type name, a letter has got lost accidentally.

* vcardcomponent: Drop declarations of non-existent functions

Either the body did not exist or it was disabled in the compile time,
thus remove it from the public API to not confuse the library users.

* vcardvalue: Remove declarations of non-existent functions

These do not have their corresponding bodies, thus drop them from
the public API, to not confuse the library users.