libical-glib: Add checker for API completeness
It adds a very simple and naive C header file parser into the generator,
with which it can read exported symbols and enums and then compares whether
all of them are included (or skipped) in the libical-glib API (.xml) files.

Configure the project with -DLIBICAL_GLIB_CHECK_API_FILES=True to enable it,
unless the project is configured with the -DLIBICAL_DEVMODE=True, which
enables it automatically.

Any new symbols should be added into the .xml files in the src/libical-glib/api/.
In case the symbol should be skipped, as not suitable for the introspection
or such, then add it into the corresponding .xml file as <skip>symbol</skip>
under a <structure/> element. Some common skipped symbols are declared
in the src/libical-glib/api/i-cal-component.xml. When the "symbol" ends
with a star (*), then it's treated as a wildcard and any symbols with such
prefix are skipped.