a small API cleanups for the 4.x (#1019)
* icalenum: Replace macros with their actual function names

These had been left for backward compatibility, remove them for
the new 4.x series.

Closes https://github.com/libical/libical/issues/1014

* libical: Declare enums as typedef-s in public headers

It allows to reference the enum with and without the `enum` keyword,
which can help with other language bindings and which can simplify
the code (by not repeating the `enum` keyword where the enum is used).