libical-glib generator fixes (#1134)
* libical-glib: Correct free function for API templates hash table in generator

It's not a GPtrArray, it's an internal structure.

Closes https://github.com/libical/libical/issues/1133

* libical-glib: Fix a memory leak in the generator code

The 'text' variable had been split and the respective values were used
in this case, thus the 'text' should be freed.

* libical-glib: Set glib program name for the generator

Then g_warning() and similar glib functions will show this program
name as a source of the message, instead of the generic "process"
string.