icalparser.c - limit iterations in in icalparser_parse (#1136)
In the case of fuzzy data with a lot of bad lines (END before BEGIN) we
limit the number of possibly bogus lines in the icalparser_get_line()
loop to keep the cpu time down.

This is mostly about keeping the fuzzer cpu limit happy.

For example: https://oss-fuzz.com/testcase-detail/5202939296153600